Bot working, added readme

+++ ChannelBot.py
@@ -1,6 +1,8 @@
 import discord
 import json
 import os
+import time
+import asyncio
 
 
 def load_token():
@@ -35,6 +37,16 @@ class ChannelBot(discord.Client):
         self.appinfo = await self.application_info()
         self.configs = load_configs()
         self.responses = load_responses()
+        try:
+            self.handler
+        except AttributeError:
+            self.handler = asyncio.ensure_future(self.delay_handler())
+
+        for guild in self.guilds:
+            if not str(guild.id) in self.configs:
+                self.configs[str(guild.id)] = self.configs["base"].copy()
+                self.configs[str(guild.id)]["channels"] = {}
+                save_configs(self.configs)
         print("Bot logged in")
 
     async def on_message(self, message):
@@ -48,11 +60,17 @@ class ChannelBot(discord.Client):
             if split[1].lower() == "help":
                 await message.channel.send(self.responses["help"])
             elif split[1].lower() == "delete":
-                await message.channel.send(self.responses["delete_help"])
+                resp = self.responses["delete_help"]
+                resp += " Current: " + str(self.configs[str(message.guild.id)]["delete_after"])
+                await message.channel.send(resp)
             elif split[1].lower() == "clear":
-                await message.channel.send(self.responses["clear_help"])
+                resp = self.responses["clear_help"]
+                resp += " Current: " + str(self.configs[str(message.guild.id)]["clear_after"])
+                await message.channel.send(resp)
             elif split[1].lower() in ["min", "minimum"]:
-                await message.channel.send(self.responses["minimum_help"])
+                resp = self.responses["minimum_help"]
+                resp += " Current: " + str(self.configs[str(message.guild.id)]["minimum_members"])
+                await message.channel.send(resp)
         elif len(split) == 3:
             if split[1].lower() == "delete":
                 mins = split[2]
@@ -122,7 +140,16 @@ class ChannelBot(discord.Client):
 
         if text_channel and len(channel.members) == 0:
             if perms.manage_channels:
-                await text_channel.delete()
+                da = self.configs[str(guild.id)]["delete_after"]
+                ca = self.configs[str(guild.id)]["clear_after"]
+                if da == 0:
+                    await text_channel.delete()
+                elif ca == 0:
+                    await text_channel.purge(limit=9999, after=text_channel.created_at)
+                elif type(da) == int or type(ca) == int:
+                    self.configs[str(guild.id)]["channels"][str(text_channel.id)] = int(time.time())
+                    save_configs(self.configs)
+
 
     async def on_voice_channel_join(self, member, channel):
         guild = channel.guild
@@ -137,7 +164,7 @@ class ChannelBot(discord.Client):
             guild.roles
         )
 
-        after_channel = discord.utils.find(
+        text_channel = discord.utils.find(
             lambda c: c.name == channel.name.lower() + "_text",
             guild.text_channels
         )
@@ -146,24 +173,81 @@ class ChannelBot(discord.Client):
             after_role = await guild.create_role(name=channel.name)
         
         if after_role:
-            if not str(guild.id) in self.configs:
-                self.configs[guild.id] = self.configs["base"]
-                save_configs(self.configs)
             create_channel = len(channel.members) >= self.configs[str(guild.id)]["minimum_members"]
-            if not after_channel and perms.manage_channels and create_channel:
+            if not text_channel and perms.manage_channels and create_channel:
                 ow = {
                     guild.default_role: discord.PermissionOverwrite(read_messages=False),
                     after_role: discord.PermissionOverwrite(read_messages=True),
                     guild.me: discord.PermissionOverwrite(read_messages=True)
                 }
-                await guild.create_text_channel(
+                text_channel = await guild.create_text_channel(
                     channel.name + "_text",
                     overwrites=ow,
                     position=channel.position,
                     category=channel.category
                 )
+            elif text_channel:
+                if str(text_channel.id) in self.configs[str(guild.id)]["channels"]:
+                    del self.configs[str(guild.id)]["channels"][str(text_channel.id)]
+                    save_configs(self.configs)
             await member.add_roles(after_role)
             print("Added Role")
+    
+    async def on_guild_join(self, guild):
+        self.configs[str(guild.id)] = self.configs["base"].copy()
+        self.configs[str(guild.id)]["channels"] = {}
+        save_configs(self.configs)
+
+    async def on_guild_remove(self, guild):
+        if str(guild.id) in self.configs:
+            del self.configs[str(guild.id)]
+            save_configs(self.configs)
+
+
+    async def delay_handler(self):
+        while True:
+            print("Handler going through stuff")
+            for gid in self.configs:
+                if gid == "base":
+                    continue
+                chid_to_delete = []
+                da = self.configs[gid]["delete_after"]
+                ca = self.configs[gid]["clear_after"]
+                for chid in self.configs[gid]["channels"]:
+                    last_visit = self.configs[gid]["channels"][chid]
+                    if type(da) != int and type(ca) != int:
+                        chid_to_delete.append(chid)
+                        continue
+
+                    elif type(da) == int and time.time() > last_visit + 60 * da:
+                        print("handler deleting stuff")
+                        channel = self.get_channel(int(chid))
+                        if channel:
+                            await channel.delete()
+                            print("Deleted Channel")
+                        else:
+                            print("Couldn't find channel")
+                        chid_to_delete.append(chid)
+
+                    elif type(ca) == int and time.time() > last_visit + 60 * ca:
+                        print("handler clearing stuff")
+                        channel = self.get_channel(int(chid))
+                        if channel:
+                            await channel.purge(limit=9999, after=channel.created_at)
+                            print("Cleared channel")
+                        else:
+                            print("Couldn't find channel")
+                        if type(da) != int:
+                            chid_to_delete.append(chid)
+
+                for chid in chid_to_delete:
+                    del self.configs[gid]["channels"][chid]
+                if len(chid_to_delete) > 0:
+                    save_configs(self.configs)
+            print("Handler still working")
+            await asyncio.sleep(30)
+
+
 
 if __name__ == "__main__":
     token = load_token()
